+++ src/math_objects/src/PolygonObject.cpp
@@ -14,15 +14,15 @@
 void PolygonObject::render(SDL_Renderer* renderer)
 {
     SDL_SetRenderDrawColor(renderer, 255, 0, 0, SDL_ALPHA_OPAQUE);
-    unsigned int i, j;
-    float x1, y1, x2, y2;
+    size_t i, j;
+    int x1, y1, x2, y2;
     j = polygon.size() - 1;
     for ( i = 0; i < polygon.size(); i++ )
     {
-        x1 = polygon.vertices[j].x * scale + dx;
-        y1 = polygon.vertices[j].y * scale + dy;
-        x2 = polygon.vertices[i].x * scale + dx;
-        y2 = polygon.vertices[i].y * scale + dy;
+        x1 = static_cast<int>(polygon.vertices[j].x * scale + dx);
+        y1 = static_cast<int>(polygon.vertices[j].y * scale + dy);
+        x2 = static_cast<int>(polygon.vertices[i].x * scale + dx);
+        y2 = static_cast<int>(polygon.vertices[i].y * scale + dy);
         SDL_RenderDrawLine(renderer, x1, y1, x2, y2);
         j = i;
     }
